feat: separate file for helper function of firebase.utils
